feat: moved util fn and save highlights individually
